+++ src/main/java/kr/hhplus/be/server/payment/domain/Payment.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.payment;
+package kr.hhplus.be.server.payment.domain;
 
 import java.time.LocalDateTime;
