+++ src/main/java/kr/hhplus/be/server/payment/domain/PaymentStatus.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.payment;
+package kr.hhplus.be.server.payment.domain;
 
 public enum PaymentStatus {
