+++ src/main/java/kr/hhplus/be/server/payment/infrastructure/persistence/SpringDataPaymentJpaRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.infrastructure.persistence.payment;
+package kr.hhplus.be.server.payment.infrastructure.persistence;
 
 import org.springframework.data.jpa.repository.JpaRepository;
